+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 0.30.0 - 2026-04-20
+### Added
+- discover IRM flat rate for configuration tree
 
 ## 0.29.3 - 2026-04-20
 ### Added